requirements.txt
requests
beautifulsoup4
psycopg2-binary
schedule
lxml
Levenshtein
pandas
numpy
tqdm
retrying